api/requirements.txt
fastapi
uvicorn
pandas
elasticsearch==8.12.1
python-dotenv